--- a/tests/fields/fields.py
+++ b/tests/fields/fields.py
@@ -6,6 +6,7 @@ import math
 import itertools
 import re
 import pymongo
+import sys
 
 from nose.plugins.skip import SkipTest
 from collections import OrderedDict
@@ -3477,6 +3478,11 @@ class FieldTest(MongoDBTestCase):
         self.assertRaises(ValidationError, user.validate)
 
     def test_email_field_unicode_user(self):
+        # Don't run this test on pypy3, which doesn't support unicode regex:
+        # https://bitbucket.org/pypy/pypy/issues/1821/regular-expression-doesnt-find-unicode
+        if sys.version_info[:2] == (3, 2):
+            raise SkipTest('unicode email addresses are not supported on PyPy 3')
+
         class User(Document):
             email = EmailField()